application form: relocate hooks to their own files
